Drink List Correction
Was able to get beers which are ready to drink or will be ready to drink
to show up in the table by changing the reference column number to 3.
Value for cells in the 5th column that are not a number still show up as
blank. This needs to be resolved.